net(kuri) : erreur de compilation avec drapeaux stricts

+++ logiciels/kuri/arbre_syntaxique/etendue_code_source.cc
@@ -41,10 +41,10 @@ void ÉtendueSourceNoeud::fusionne(ÉtendueSourceNoeud autre)
 {
     auto const pos = position_lexeme(*lexème);
     ÉtendueSourceNoeud résultat;
-    résultat.ligne_début = pos.index_ligne;
-    résultat.colonne_début = pos.pos;
-    résultat.ligne_fin = pos.index_ligne;
-    résultat.colonne_fin = pos.pos + lexème->chaine.taille();
+    résultat.ligne_début = int32_t(pos.index_ligne);
+    résultat.colonne_début = int32_t(pos.pos);
+    résultat.ligne_fin = int32_t(pos.index_ligne);
+    résultat.colonne_fin = int32_t(pos.pos + lexème->chaine.taille());
     return résultat;
 }
